extend input logic classes to the game scene; change folder structure
