+++ client/src/main.ts
@@ -1,5 +1,5 @@
 import { Boot } from './scenes/Boot';
-import { Game as MainGame } from './scenes/Game';
+import { Game as MainGame } from './scenes/Game/Game';
 import { GameSetup as MainGameSetup } from './scenes/GameSetup/GameSetup';
 import { GameOver } from './scenes/GameOver';
 import { MainMenu } from './scenes/MainMenu';
