+++ client/src/modalities/VoiceInput.ts
@@ -1,0 +1,5 @@
+/**
+ * generic methods to handle voice inputs
+ * @abstract
+ */
+export abstract class VoiceInput {}
